infra/local-api/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.34.0
asyncpg==0.30.0
python-dotenv==1.0.1
httpx==0.28.1
